+++ package/__tests__/index.js
@@ -1,4 +1,5 @@
 const index = require('../index')
+const { generateWebpackConfig } = require("../index");
 
 describe('index', () => {
   test('exports webpack-merge v5 functions', () => {
@@ -18,5 +19,28 @@ describe('index', () => {
 
     expect(webpackConfig2).not.toHaveProperty('newKey')
     expect(webpackConfig2.output.path).not.toEqual('new value')
+  })
+
+  test('webpackConfig merges extra config', () => {
+    const { generateWebpackConfig } = require('../index')
+
+    const webpackConfig = generateWebpackConfig({
+       newKey: 'new value',
+       output: {
+         path: 'new path'
+      }
+    })
+
+    expect(webpackConfig).toHaveProperty('newKey', 'new value')
+    expect(webpackConfig).toHaveProperty('output.path', 'new path')
+    expect(webpackConfig).toHaveProperty('output.publicPath', '/packs/')
+  })
+
+  test('webpackConfig errors if multiple configs are provided', () => {
+    const { generateWebpackConfig } = require('../index')
+
+    expect(() => generateWebpackConfig({}, {})).toThrow(
+      'use webpack-merge to merge configs before passing them to Shakapacker'
+    )
   })
 })