+++ spec/dummy/config/webpack/commonWebpackConfig.js
@@ -13,6 +13,6 @@ const ignoreWarningsConfig = {
 };
 // Copy the object using merge b/c the baseClientWebpackConfig and commonOptions are mutable globals
 // const commonWebpackConfig = () => (merge({}, baseClientWebpackConfig, commonOptions))
-const commonWebpackConfig = () => (merge({}, generateWebpackConfig(), commonOptions, ignoreWarningsConfig))
+const commonWebpackConfig = () => generateWebpackConfig(merge(commonOptions, ignoreWarningsConfig))
 
 module.exports = commonWebpackConfig